Samsung version of android come with a fancy mode called "Multi Window", where you can split the screen and run two applications simultaneously.
This requires a new tag in the manifest under the <application> section

<uses-library required="false" name="com.sec.android.app.multiwindow" />

Seems like it should be harmless to include this in the template as it is
not required by default.

There are other required tags to enable multi-window support, but buildozer
allows to update the template dynamically for the rest of the tags, so
those things don't need to go into this template.