unpack into subdirectories __kivy_root__ of private and external storage.
Since unpacking performs first a recursiveDelete, everything stored in an
unpack root is lost after an app update.  In order for e.g. a database on
private storage to persist across app updates it was necessary to make
the private unpack root a subdirectory of private storage rather than
the top level.