Support version in --requirements
In the old toolchain, we were able to specify a version in the requirements (-m "kivy==master,zeroconf==1.9")
This allow to replicate the same behavior.

I'm unsure about the distribution matching with version, so i did made a change in order to compare distribution only on names, not name+version if exists. Any advises?