Empty: Adding .gitkeep in bootstraps/empty/build
When experimenting around new recipes for python3, I personally needed empty to get something built.
The reason is that almost every boostrap needs python2, which conflicts with python3.
While using empty I noticed that the build directory is missing.

Using .gitkeep is not documented in git, but widely used as far as I read on the internet.