Add support for minimum Android API.
This allows the compilation to use a different version for compiling
while targetting a higher API version.

Typically, you may need to use API 23 for Java, but want to run on API
19. If you compile with API 23, when running on older device (let's say
4.4.2), you'll end up with a symbol not found "srand", "signal", etc.

It is important to set the Android API minimum version when compiling in
order to avoid theses symbol issues.

CF: https://stackoverflow.com/a/41079462/69877