Fixes and clarifies services auto-restart
1) Fixes services auto-restart introduced in #643:
`onTaskRemoved()` stops the service gracefully using `stopSelf()` when
it gets killed. That way `onDestroy()` gets triggered and service will
restart under different process if auto-restart is enabled.

2) Documents process behavior and auto-restart usage:
Emphasis that services are running in a dedicated processes, different
from the app process, which is not the Android default behavior.
Adds a code snippet.