Conditional recipe build fixes #1382
This is a first iteration addressing recipes conditional build.
In this version all modified recipes are build in batch with no
dependency resolution whatsoever.
Also recipes are only built against `python3crystax` target and
`sdl2` bootstrap.
The list of broken recipes is not fully known, but this script will help
finding out. If a recipe is known to be broken, it should be added in
the `BROKEN_RECIPES` set to be skipped until it gets fixed.