Conditional builds falls back on Python2 if needed
Falls back to python2 when tested recipes are python3crystax compatible.
This is for instance the case on #1428 where the recipes depend only on
python2, hence couldn't be automatically tested.

Also updated `vispy` recipe to handle the version string properly,
pin it and clean comments.
This recipe is a demo case scenario for the conditional build falling
back on python2 since `vispy` is not yet compatible with python3crystax.

Last, removes the debug flag from the test apps used by the CI as it
produces too many logs and halts Travis.