[protobuf_cpp] Updated to Protobuf 3.5.1 + fixed protoc issue
Protobuf has been updated to 3.5.1, it's not the last version (3.6.1)
because last version cause trouble during compilation and needs more
work.

This patch also fixes the issue with hardcoded local path for protoc which
could not work. Protobuf needs a version of "protoc" (its compiler, used
to produce .py file from .proto files) compiled for the host platform.
Official binaries for protoc are available, the recipe will download the
one suitable for host and use it.
If the platform is not supported, building may still work if the right "protoc"
is available in search path.