Optimize Dockerfile
Prioritize SDK and NDK downloads before everything because once downloaded we don't care about redownloading unless the version is changed and layer cache works this way more efficiently. Redirect Licenses to /dev/null, move more important layers up. Move '&&' operators to the front - Docker sometimes, (personally I've encountered it only with apt and old coreutils), is very critical of the amount of spaces *after* the operator but does not care what's in front of it that much.