Adapt service_only's bootstrap c/java code to new python build system
Also add some files because we need it because of the changes made in PythonService.java. Those files add the classes PythonActivity,  AssetExtract and ResourceManager (those two lasts are pure copies from the ones in sdl2 bootstrap).

Note: All this work is based on previous work done on sdl2 bootstrap