Add pyzbar recipe and remove duplicated flags for libzbar
The libzbar recipe contains some flags already set in base class method of get_recipe_env, so removed. Also grants python3 compatibility for libzbar because, despite that the zbar recipe isn't compatible with python3 (the project is stucked in python2 plus it seems unmaintained)...but there are alternatives which can replace it, so we add the pyzbar recipe which works for both versions of python so python3 will have an equivalent to the python2's zbar. Also moved libraries from LDFLAGS to LIBS.