Suppress almost all logs for ci builds (fix travis limit of 4mb of logs)
In order to not exceed the travis limitation of 4MB of logs...this will allow us to add more tests without facing this limitation or update the android build tools. It must be said that if any error happens the last 1000 lines of the log will be printed, so we know where is the problem. This will have the advantage that we will focus the problem more quickly but it maybe some cases where the error happens before of the printed lines...if that is the case we should disable temporary this feature, probably with some tests, and rerun the travis tests to get the full log.