Allow patching from any folder + fix pygame components issues
This is needed to properly apply the patches for the pygame bootstrap components recipe, because we must apply the patches from build_container_dir and the current patching method only allow us to do that from the build directory (which not exists in pygame components because we delete it). Also fix a typo error for pygame components.