Speed up Docker chown via COPY parameter
For some reason running the `chown` command explicitly is hanging on
Docker. Using the `--chown` parameter in the `COPY` command is faster.
However that new method has a couple of drawbacks:
1) it was introduced in a recent Docker version 17+
2) it doesn't accept variables as argument, only hardcoded values

Unfortunately we still need to use explicit "shell" `chown` for the
`ANDROID_SDK_HOME` as this directory must still be writable. Error was:
```
A problem occurred configuring root project 'bdisttest_python2'.
> Failed to install the following SDK components:
      build-tools;27.0.3 Android SDK Build-Tools 27.0.3
  The SDK directory is not writable (/opt/android/android-sdk)
```