Auto deploys to PyPI using Travis on tags
Uses PyPI `__token__` and Travis auto deployment to publish to PyPI
upon tagging.
https://pypi.org/help/#apitoken
Travis uses the not so well documented `PYPI_PASSWORD` environment
variable to store the password.
https://github.com/travis-ci/dpl/blob/v1.10.15/lib/dpl/provider/pypi.rb#L12
The variable was set with a valid token prior this change.
https://travis-ci.org/github/kivy/python-for-android/settings
This is more secure than user and password as tokens are generated with
uploaded only permissions and can be scoped to a given project.