Update PythonActivity.java to support webview file upload
It may seem scary, but it's useful if we want to use python to make a twitter-like app. After all, we need to upload image than bare text.

But take your time, I'm not hurrying with this pull. Because I'm working on a project with the use of react.js and python for android.

I may discovery more to improve the usability of `python for android`. We'll see.