Resize webview when keyboard is shown
When the adjust mode is unspecified, Android will not resize the webview
when showing the on screen keyboard to keep the input element on screen.
Explicitly instruct it to use `adjustResize` mode. Possibly this is
because the webview bootstrap uses an `AbsoluteLayout` as the top level
widget. See
https://developer.android.com/guide/topics/manifest/activity-element#wsoft
for details.