Return proper HTTP responses in webview testapp
Current Flask throws an exception if a view function returns `None`:

```
python  : TypeError: The view function for 'vibrate' did not return a
valid response. The function either returned None or ended without a
return statement.
```

Maybe that was different in the past, but instead a proper response can
be returned. An empty 200 would be fine, but a 204 No Content response
is used since it better represents what's happening.