Support WorkManager tasks
The androidx WorkManager component allows simple handling of tasks in
Android apps. Furthermore, in Android 12, WorkManager tasks are the
preferred way to run foreground services when the app is in the
background.

Since the python interpreter integration only allows running one python
interpreter per process, this uses a `RemoteListenableWorker` to run the
tasks on a service in a separate process. Unfortunately, that's only
available since the work-multiprocess 2.6.0 release, which requires SDK
30.

Co-authored-by: Robert Niederreiter <office@squarewave.at>
Co-authored-by: Philipp Auersperg <phil@bluedynamics.com>