android/activity: Add Application.ActivityLifecycleCallbacks helpers
The `Application.ActivityLifecycleCallbacks` interface is used to
register a class that can receive callbacks on Activity lifecycle
changes. Add a `PythonJavaClass` implementing the interface and helper
functions to register python callbacks with it. This can be used to
perform actions in the python app when the activity changes lifecycle
states.