:construction_worker: Publish to DockerHub
Our public DockerHub Image was 3 years old as the automatic build became
a pro feature.
With this change we're now publishing to DockerHub from the CI on push
to the develop branch.

Authentication to DockerHub is done using personal access tokens
configured using `andremiras` username, but this can be updated
anytime with another token refs:
- https://app.docker.com/settings/personal-access-tokens
- https://github.com/kivy/python-for-android/settings/secrets/actions

We use a dedicated docker.yml workflow file rather than building on top
of the existing push.yml one to keep things clean and separated. In the
future we may add multi architecture support.

Updating the public image also fixes the issue where the sh logger was
too verbose leading to huge and unreadable CI logs.